requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
PyYAML
numpy
soundfile
torch
nemo_toolkit
omegaconf
pytest
